chore :: 랜딩페이지 스타일 정리
