+++ components/landing/service/style.ts
@@ -1,12 +1,10 @@
-import React from 'react';
 import styled from '@emotion/styled';
-import { THEME } from '../../styles/theme';
-import z from '../../public/zz.svg';
+import { THEME } from '../../../styles/theme';
+import z from '../../../public/zz.svg';
 import { keyframes } from '@emotion/react';
 
-
 // 서비스 섹션 전체 래퍼
-const SectionWrapper = styled.section`
+export const SectionWrapper = styled.section`
   width: 100vw;
   min-height: 100vh;
   background: #d5e3e2;
@@ -22,20 +20,14 @@ const SectionWrapper = styled.section`
   background-repeat: no-repeat;
 `;
 
-const ArrowSvg = () => (
-  <svg width="64" height="40" viewBox="0 0 64 40" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path d="M32 8V32M32 32L16 16M32 32L48 16" stroke={THEME.colors.Primary.Normal} strokeWidth="4" strokeLinecap="round" strokeLinejoin="round"/>
-  </svg>
-);
-
-const bounceArrow = keyframes`
+export const bounceArrow = keyframes`
   0% { transform: translateX(-50%) translateY(0); }
   50% { transform: translateX(-50%) translateY(24px); }
   100% { transform: translateX(-50%) translateY(0); }
 `;
 
 // 타이틀/서브타이틀 래퍼
-const TitleBlock = styled.div`
+export const TitleBlock = styled.div`
   width: 100%;
   max-width: 900px;
   margin: 0 auto 48px auto;
@@ -44,7 +36,7 @@ const TitleBlock = styled.div`
   align-items: center;
 `;
 
-const Title = styled.h2`
+export const Title = styled.h2`
   font-family: 'Pretendard Variable', 'SF Pro', ${THEME.typography.fontFamily.join(', ')};
   font-size: 48px;
   font-weight: 700;
@@ -54,7 +46,7 @@ const Title = styled.h2`
   text-align: center;
 `;
 
-const Subtitle = styled.p`
+export const Subtitle = styled.p`
   font-family: 'Pretendard Variable', 'SF Pro', ${THEME.typography.fontFamily.join(', ')};
   font-size: 16px;
   font-weight: 400;
@@ -64,7 +56,7 @@ const Subtitle = styled.p`
   padding-bottom: 80px;
 `;
 
-const DownArrow = styled.div`
+export const DownArrow = styled.div`
   position: absolute;
   left: 50%;
   bottom: 40px;
@@ -80,7 +72,7 @@ const DownArrow = styled.div`
 `;
 
 // 카드 리스트 래퍼
-const CardRow = styled.div`
+export const CardRow = styled.div`
   width: 100%;
   max-width: 1520px;
   display: flex;
@@ -91,7 +83,7 @@ const CardRow = styled.div`
 `;
 
 
-const Card = styled.div`
+export const Card = styled.div`
   flex: 1 1 0;
   min-width: 280px;
   max-width: 360px;
@@ -109,7 +101,7 @@ const Card = styled.div`
   backdrop-filter: blur(8px);
 `;
 
-const CardTitleRow = styled.div`
+export const CardTitleRow = styled.div`
   width: 100%;
   display: flex;
   flex-direction: row;
@@ -118,7 +110,7 @@ const CardTitleRow = styled.div`
   margin-bottom: 12px;
 `;
 
-const CardTitle = styled.h3`
+export const CardTitle = styled.h3`
   font-family: 'Pretendard Variable', 'SF Pro', ${THEME.typography.fontFamily.join(', ')};
   font-size: 24px;
   font-weight: 600;
@@ -126,7 +118,7 @@ const CardTitle = styled.h3`
   line-height: 1.1;
 `;
 
-const CardIcon = styled.span`
+export const CardIcon = styled.span`
   display: flex;
   align-items: center;
   justify-content: center;
@@ -139,7 +131,7 @@ const CardIcon = styled.span`
   font-weight: 700;
 `;
 
-const CardDesc = styled.p`
+export const CardDesc = styled.p`
   font-family: 'Pretendard Variable', 'SF Pro', ${THEME.typography.fontFamily.join(', ')};
   font-size: 16px;
   font-weight: 400;
@@ -148,48 +140,3 @@ const CardDesc = styled.p`
   margin: 0;
   white-space: pre-line;
 `;
-
-// 카드 데이터
-const cards = [
-  {
-    title: '해시태그 추천',
-    desc: '키워드 기반 AI 추천으로\n트렌드에 맞는 해시태그를 추천해 줍니다.',
-  },
-  {
-    title: '제목 분석',
-    desc: '썸네일과 제목을 분석하여\n클릭률을 높이는 최적의 조합을 제안해 줍니다.',
-  },
-  {
-    title: '콘텐츠 캘린더',
-    desc: '요일별 맞춤 일정 추천으로\n규칙적인 콘텐츠 업로드 및 관리를 돕습니다.',
-  },
-  {
-    title: 'AI 콘텐츠 피드백',
-    desc: '개선점과 반응 예측을 통해\n더 나은 양질의 콘텐츠를 제작하도록 돕습니다.',
-  },
-];
-
-const ServiceSection: React.FC = () => (
-  <SectionWrapper>
-    <TitleBlock>
-      <Title>함께 성장하는 도구들</Title>
-      <Subtitle>씨앗부터 만개까지, 여러분의 콘텐츠 제작 여정을 함께하는 여섯 가지 강력한 도구입니다.</Subtitle>
-    </TitleBlock>
-    <CardRow>
-      {cards.map((card, idx) => (
-        <Card key={card.title}>
-          <CardTitleRow>
-            <CardTitle>{card.title}</CardTitle>
-            <CardIcon>→</CardIcon>
-          </CardTitleRow>
-          <CardDesc>{card.desc}</CardDesc>
-        </Card>
-      ))}
-    </CardRow>
-    <DownArrow>
-      <ArrowSvg />
-    </DownArrow>
-  </SectionWrapper>
-);
-
-export default ServiceSection;
